// projects/cpm-app/src/app/shared/model/i-device-location-access-component-data.ts
export interface IDeviceLocationAccessComponentData {
  SlotNumber: number;
  BinNumber: number;
  ShelfNumber: number;
  DeviceLocationTypeId: string;
  DeviceId: number;
  DeviceLocationId: number;
  DeviceLocationAccessQuantity: number;
  DeviceLocationAccessUnits: string;
  ItemTradeName: string;
  ItemGenericNameFormatted: string;
}